requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-graphql-jwt==0.4.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
